Make the code PEP8 compliant.
This commit tries to make the code more PEP8 compliant by fixing most of
the reports generated by the flake8 tools. Contrary to PEP8, I set the
maximum line length to a more reasonable value of 120 characters.

There are some whitespace reportings left, especially regarding the file
chord_table.py. However, this is intended to keep the chords tables
readable and totally consistent with PEP20 (The Zen of Python).

This work is preparing the port to Python 3 as Python 2 is close to EOL.

The folling changes have been made:

 - Cleanup unnecessary blank lines
 - Add blank lines where necessary
 - Use 'is' and  'is not' operator for comparisons to None
 - Split the use of multiple statements on one line (E701)
 - Remove semicola from statements (E703)
 - Remove unnecessary whitespaces (E201, E202)
 - Add whitespaces where necessary
 - Remove some useless inline comments
 - Replace '<>' with '!=' (W603)
 - Remove redundant backslash (E502)
 - Remove mixed indention with tabstops and whitespace
 - Sort module imports according to PEP8
 - Remove line break before binary operator (W503)
 - Fix ambigous variable naming (E741)
 - Fix usage of bare 'except' (E722)
 - Replace escape sequences with corresponding utf-8 symbol (W605)
 - Improve indention and alignment to increase readability
 - Fix wrong comment indention
 - Align line continuations with visual indent
 - Reduce length of some lines to keep em < 120 chars